+++ cadence-sensor/CadenceCalculator.cpp
@@ -12,7 +12,8 @@ CadenceCalculator::CadenceCalculator(void)
   : _state{ AppState_t::CALCULATE },
     _prevCumlativeCranks{ 0 },
     _prevLastWheelEventTime{ 0 },
-    _lastCalculateTime{ 0 } {}
+    _lastCalculateTime{ 0 },
+    _sessionCranks{0} {}
 
 CadenceCalculator::~CadenceCalculator(void) {}
 
@@ -23,6 +24,7 @@ bool CadenceCalculator::initialize(void) {
   _prevCumlativeCranks = 0;
   _prevLastWheelEventTime = 0;
   _lastCalculateTime = 0;
+  _sessionCranks = 0;
 
   return true;
 }
@@ -53,6 +55,8 @@ void CadenceCalculator::_calculate(void) {
         deltaRotations += 0xFFFF;
       }
 
+      _sessionCranks = _sessionCranks + deltaRotations;
+
       int32_t timeDelta = blackboard.ble.lastWheelEventTime - _prevLastWheelEventTime;
       if (timeDelta < 0) {
         // Roll over
@@ -62,9 +66,13 @@ void CadenceCalculator::_calculate(void) {
 
       if ((timeDelta != 0) && (deltaRotations != 0)) {
         // Convert event time delta to a time in minutes
-        float const timeMins = static_cast<float>(timeDelta) / SENSOR_TIME_RESOLUTION / SECONDS_PER_MINUTE;
+        float_t const timeMins = static_cast<float>(timeDelta) / SENSOR_TIME_RESOLUTION / SECONDS_PER_MINUTE;
         // Calculate new RPM
         blackboard.cadence.cadence = static_cast<uint16_t>(static_cast<float>(deltaRotations) / timeMins);
+
+        float_t const distanceMiles = static_cast<float_t>(_sessionCranks) * WHEEL_CIRCUMFERENCE_INCHES * INCHES_TO_MILES;
+
+        blackboard.cadence.distance = distanceMiles;
         _lastCalculateTime = millis();
         Log.noticeln("Cadence %d", blackboard.cadence.cadence);
         // Save latest data